twitter-app/requirements.txt
elasticsearch==7.5.1
email_validator==1.1.2
flask_bootstrap==3.3.7.1
flask_httpauth==1.0.2
flask_login==0.5.0
flask_migrate==2.5.3
flask_moment==0.5.2
flask_paginate==1.0.2
flask_sqlalchemy==2.4.4
flask_wtf==0.14.3
flask==1.1.2
flasgger==0.9.5
python-dotenv==0.15.0
wtforms==2.3.3